Make the 'Run' button switch to a 'stop' button when pipeline is running. This should make the UI a bit more intuitive.
Achieve this by swapping the 'run' tool in the toolbar with a 'stop'
tool. Swapping tools is needed because although the pixmap for a tool
can be set (with wxToolBar::SetToolNormalBitmap), the label can't.
The currently hidden tool is stored in hiddenStartStopTool.  When
constructing the tools, use a 'add both to the toolbar then remove
'stop' tool' trick, so that wxToolBar::AddTool() can be used simply as
a constructor for tools.