count frames and add timestamp in ComponentInputCameraUSB by keeping count of the frames, and creating a timestamp when the image is read.
This makes FrameNumber and TimeStamp usable when the input is an USB camera.

Note that the other two camera components keep a count of the frames
as well. I am not sure whether reading the frame number from the
camera should work (it allways gives -1 for our camera/driver (v4l2)).

This change was actually done by Gabor Vasarhelyi, I am merging it
because he doesn't have commit rights (yet).